requirements.txt
requests
lxml
pytesseract
Pillow
cssselect


